gamebackend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.2
python-dotenv==1.0.0 